pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geospatial-vit"
version = "0.1.0"
description = "Vision Transformer for Geospatial Drone Imagery"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "pillow>=10.0.0",
    "albumentations>=1.3.0",
    "timm>=0.9.0",
    "einops>=0.6.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
]
